+++ lib/users/types.ts
@@ -1,22 +1,22 @@
 export declare type User = {
-    user_id?: number;
-    org_id?: number;
-    name?: string;
-    email?: string;
-    phone?: string;
-    permissions?: string;
-    team_access?: boolean;
-    role?: string;
-    status?: string;
-    avatar?: string;
-    time_zone?: string;
-    location?: string;
-    team?: string;
-    is_verified?: boolean;
-    is_unsubscribed?: boolean;
-    last_login_at?: Date;
-    last_login_ip?: string;
-    uid?: string;
-    updated_at?: Date;
-    created_at?: Date;
-}
+  user_id?: number;
+  org_id?: number;
+  name?: string;
+  email?: string;
+  phone?: string;
+  permissions?: string;
+  team_access?: boolean;
+  role?: string;
+  status?: string;
+  avatar?: string;
+  time_zone?: string;
+  location?: string;
+  team?: string;
+  is_verified?: boolean;
+  is_unsubscribed?: boolean;
+  last_login_at?: Date;
+  last_login_ip?: string;
+  uid?: string;
+  updated_at?: Date;
+  created_at?: Date;
+};
